refactor: merge rpc handler files into single rpc_table.go
Collapse 5 separate rpc_*.go files into one rpc_table.go (752 lines).
Use anonymous struct params in closures instead of 29 named type defs.
Net reduction: -176 lines vs master.
